move Feed component
