+++ client/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Feed from './components/Feed'
+import Feed from './components/Feed/Feed'
 
 function App () {
   return (
